requirements.txt
microfs==1.4.5
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyserial==3.5
